+++ src/trpc/server.ts
@@ -1,20 +1,12 @@
-import { inferAsyncReturnType, initTRPC } from "@trpc/server";
-import {
-  CreateExpressContextOptions,
-  createExpressMiddleware,
-} from "@trpc/server/adapters/express";
+import { initTRPC } from "@trpc/server";
+import { createExpressMiddleware } from "@trpc/server/adapters/express";
+import { Context, createContext } from "./fetch-context.ts";
+// import { Context, createContext } from "./express-context.ts";
 import { z } from "zod";
-
-// created for each request
-const createContext = ({
-  req,
-  res,
-}: CreateExpressContextOptions) => ({}); // no context
-type Context = inferAsyncReturnType<typeof createContext>;
 
 const t = initTRPC.context<Context>().create();
 
-const appRouter = t.router({
+export const appRouter = t.router({
   hello: t.procedure
     .input(z.string())
     .query((req) => {